+++ rentfetch-sync.php
@@ -3,7 +3,7 @@
 	Plugin Name: Rent Fetch Sync
 	Plugin URI: https://github.com/jonschr/rentfetch-sync
 	Description: An addon for Rent Fetch that syncs properties, floorplans, and units 
-	Version: 0.4.2
+	Version: 0.4.3
 	Author: Brindle Digital
 	Author URI: https://www.brindledigital.com/
 
@@ -24,7 +24,7 @@ if ( !defined( 'ABSPATH' ) ) {
 }
 
 // Define the version of the plugin
-define ( 'RENTFETCHSYNC_VERSION', '0.4.2' );
+define ( 'RENTFETCHSYNC_VERSION', '0.4.3' );
 
 // Plugin directory
 define( 'RENTFETCHSYNC_DIR', plugin_dir_path( __FILE__ ) );
@@ -115,10 +115,10 @@ function rfs_start_sync_single_property() {
 	// RealPage doesn't have any property information or photos
 	
 	// define what to sync
-	rfs_sync_single_property( $property_id = 'p1585665', $integration = 'yardi' );
+	rfs_sync_single_property( $property_id = '4818643', $integration = 'realpage' );
 	
 }
-// add_action( 'wp_footer', 'rfs_start_sync_single_property' );
+add_action( 'wp_footer', 'rfs_start_sync_single_property' );
 
 // Load Plugin Update Checker.
 require RENTFETCHSYNC_DIR . 'vendor/plugin-update-checker/plugin-update-checker.php';